Feat/v1 name registry

+++ Circles.Pathfinder/Edges/FlowEdge.cs
@@ -1,5 +1,3 @@
-using System.Numerics;
-
 namespace Circles.Pathfinder.Edges;
 
 /// <summary>
@@ -7,11 +5,11 @@ namespace Circles.Pathfinder.Edges;
 /// </summary>
 public class FlowEdge : CapacityEdge
 {
-    public BigInteger CurrentCapacity { get; set; }
-    public BigInteger Flow { get; set; }
+    public long CurrentCapacity { get; set; }
+    public long Flow { get; set; }
     public FlowEdge? ReverseEdge { get; set; }
 
-    public FlowEdge(string from, string to, string token, BigInteger initialCapacity)
+    public FlowEdge(string from, string to, string token, long initialCapacity)
         : base(from, to, token, initialCapacity)
     {
         CurrentCapacity = initialCapacity;
